[dev-launcher] refactor modal component
