+++ packages/expo-dev-launcher/bundle/components/DevServerExplainerModal.tsx
@@ -1,9 +1,11 @@
 import { Text, Spacer, View } from 'expo-dev-client-components';
 import * as React from 'react';
 
+import { BaseModal } from './BaseModal';
+
 export function DevServerExplainerModal() {
   return (
-    <View>
+    <BaseModal title="Development servers">
       <Spacer.Vertical size="small" />
       <Text size="medium">Start a local development server with:</Text>
       <Spacer.Vertical size="small" />
@@ -18,6 +20,6 @@ export function DevServerExplainerModal() {
       <Text>
         Alternatively, open the Camera app and scan the QR code that appears in your terminal
       </Text>
-    </View>
+    </BaseModal>
   );
 }
