+++ packages/expo-dev-launcher/bundle/components/LoadAppErrorModal.tsx
@@ -1,10 +1,12 @@
-import { View, Text } from 'expo-dev-client-components';
+import { Text } from 'expo-dev-client-components';
 import * as React from 'react';
+
+import { BaseModal } from './BaseModal';
 
 export function LoadAppErrorModal({ message }) {
   return (
-    <View>
+    <BaseModal title="Error loading app">
       <Text>{message}</Text>
-    </View>
+    </BaseModal>
   );
 }
